Move coverage functionality from old-style controller to new one
